requirements.txt
mysql-connector-python==8.0.33
cryptography==41.0.3
bcrypt==4.1.2
requests==2.31.0